server/requirements.txt
# Note: This project requires Python 3.12 or earlier due to mlx-audio dependencies
python-dotenv
fastapi[all]
uvicorn
opencv-python>=4.12.0.88
mlx-lm==0.24.0
mlx-audio
pipecat-ai[openai,deepgram,rime,silero,mlx-whisper]>=0.0.76
aiortc
nltk
resemblyzer>=0.1.3
beautifulsoup4>=4.12.0
html2text>=2020.1.16
pytz>=2024.1
aiosqlite>=0.20.0

# MCP Proxy for tool integration with LM Studio
mcpo>=0.0.17

# browser tools
trafilatura
tools

# --- Sherpa-ONNX (new) ---
numpy>=1.26
sherpa-onnx>=1.12.8
onnxruntime>=1.22.1
huggingface_hub>=0.23.0

# --- DuckDuckGo Search ---
ddgs>=9.5.2

# --- Mem0 and HuggingFace Embeddings ---
mem0ai>=0.1.26
sentence-transformers>=5.1.0